moved np.floats to floats, added readme  (#1)
* Added readme, updated all np.float  to float

* added exini flags to version and readme

* version change

+++ sklearn/ensemble/_bagging.py
@@ -311,7 +311,7 @@ class BaseBagging(BaseEnsemble, metaclass=ABCMeta):
         # Validate max_features
         if isinstance(self.max_features, numbers.Integral):
             max_features = self.max_features
-        elif isinstance(self.max_features, np.float):
+        elif isinstance(self.max_features, float):
             max_features = self.max_features * self.n_features_
         else:
             raise ValueError("max_features must be int or float")
